+++ packages/common/src/node.js
@@ -24,6 +24,7 @@ const {
 
 const {
   callbackify,
+  debounce,
   forEach,
   lruCache,
   promisify,
@@ -53,6 +54,7 @@ module.exports = {
 
   utils,
     callbackify,
+    debounce,
     forEach,
     lruCache,
     promisify,