+++ packages/common/src/utils/debounce.js
@@ -1,0 +1,11 @@
+const debounce = (fn, waitMs = 1000) => {
+  let timerId
+  return (...args) => {
+    if (timerId) {
+      clearTimeout(timerId)
+    }
+    timerId = setTimeout(() => fn(...args), waitMs)
+  }
+}
+
+module.exports = debounce