+++ packages/common/src/utils/index.js
@@ -1,4 +1,5 @@
 const callbackify = require('./callbackify')
+const debounce = require('./debounce')
 const forEach = require('./for-each')
 const lruCache = require('./lru-cache')
 const promisify = require('./promisify')
@@ -6,6 +7,7 @@ const rateLimit = require('./rate-limit')
 
 module.exports = {
   callbackify,
+  debounce,
   forEach,
   lruCache,
   promisify,